Fix Kismet log saving, remove passive
Tested Kismet log saving and noticed that it was only copying logs to sdcard when using GPS support (which is mainly what I used).  Moved the log saving to a sperate function.

Also added -t to kismet so that kalimenu knows which files to look for when zipping that way we don't have to zip the entire folder, just that sessions.

Testing so far shows that its working.

Signed-off-by: binkybear <binkybear@nethunter.com>