Add busybox64 bit, nexus 9 testing, Image.gz-dtb
Added Nexus 9 install script.  Possible issues with installer as I'm not sure if I have right /system path and the Nexus 9 spits out Image.gz-dtb which is different kernel file also.

Added support for that in androidmenu.sh but we are creating a boot.img anyways so it's more for future arm64.

Signed-off-by: binkybear <binkybear@nethunter.com>