Change paths nethunterbuild.sh
Since we have to git-clone to even run nethunterbuild.sh, we can't assume that they will download only that file and run it from ~/arm-stuff.

We have to assume that they will be running it from whatever folder they git cloned, so we should use `pwd`.

We can use the toolchain to detect if we have ever done a dependency check.

Signed-off-by: binkybear <binkybear@nethunter.com>