Initial upload
Initially this was based on the fattire branch of Nethunter app.  However, because of some issues preventing execution in app this is a temporary fix to use the new install method with old chroot install method.

This is mainly an execuse to use the installer which works better then old.  It auto installs busybox/supersu and allows for app selection.

It doesn't touch the ramdisk when installing except for with Marshmallow.

Signed-off-by: binkybear <binkybear@nethunter.com>