[Nethunter FS] Add Docker build script/instruction
Added instructions to README to build a rootfs from Docker.  Should be a simple three commands to get up and running to build a rootfs.  This has the added benefit of everyone being on the same page for the build environment and hopefully reduce problems when building from other distros.

Tested on OSX and build was completed.  Did not test finished build on Android device but it appears to have completed.

Signed-off-by: binkybear <binkybear@nethunter.com>